Move the logic that verifies if a Qgis Layer is LADM to DbLadmLayerTester implementations
