+++ asistente_ladm_col/db_support/db_ladm_layer_tester.py
@@ -3,7 +3,7 @@
 /***************************************************************************
                               Asistente LADM_COL
                              --------------------
-        begin                : 2019-02-21
+        begin                : 2019-03-15
         git sha              : :%H$
         copyright            : (C) 2019 by Yesid Polanía (BSF Swissphoto)
         email                : yesidpol.3@gmail.com
@@ -16,38 +16,12 @@
  *                                                                         *
  ***************************************************************************/
 """
-from qgis.PyQt.QtCore import QObject
-from .enum_action_type import EnumActionType
 
 
-class DbAdmin(QObject):
+class DbLadmLayerTester:
 
     def __init__(self):
-        self._mode = None
+        pass
 
-    def get_id(self):
-        raise Exception('unimplemented method')
-
-    def get_name(self):
-        raise Exception('unimplemented method')
-
-    def get_config_panel(self):
-        raise Exception('unimplemented method')
-
-    def get_model_baker_tool_name(self):
-        raise Exception('unimplemented method')
-
-    def get_db_connector(self, parameters):
-        raise Exception('unimplemented method')
-
-    def get_schema_import_configuration(self, params):
-        raise Exception('unimplemented method')
-
-    def get_import_configuration(self, params):
-        raise Exception('unimplemented method')
-
-    def get_export_configuration(self, params):
-        raise Exception('unimplemented method')
-
-    def get_ladm_layer_tester(self):
+    def is_ladm_layer(self, layer, db):
         raise NotImplementedError
