+++ asistente_ladm_col/db_support/none_ladm_layer_tester.py
@@ -1,0 +1,8 @@
+from .db_ladm_layer_tester import DbLadmLayerTester
+from qgis.core import (QgsDataSourceUri)
+
+
+class NoneLadmLayerTester(DbLadmLayerTester):
+
+    def is_ladm_layer(self, layer, db):
+        return False
